pyproject.toml
[tool.poetry]
name = "django-generic-helpers"
version = "1.1.0"
description = "A small toolset that helps you to work with Django's generic relations"
authors = [
    "Mikhail Porokhovnichenko <marazmiki@gmail.com>"
]
license = "MIT"
readme = "README.rst"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.1",
    "Framework :: Django :: 3.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10"
]

homepage = "https://github.com/marazmiki/django-generic-helpers"
repository = "https://github.com/marazmiki/django-generic-helpers"
keywords = ["django", "framework", "relations", "generic", "generic relations"]
packages = [
    {include = "generic_helpers", from = "src"},
]


[tool.poetry.dependencies]
python = "^3.6 || ^3.7 || ^3.8 || ^3.9 || ^3.10"

[tool.poetry.dev-dependencies]
tox = "^3"
tox-pyenv = "*"
pytest = "^6.0"
pytest-django = "^4"
pytest-env = "^0.6.2"
pytest-isort = "^1"
pytest-pythonpath = "^0.7.3"
pytest-flake8 = "^1.0"
pytest-cov = "^2.7"
bump2version = "^0.5.11"
flake8 = "^3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
